Refactor for App Store readiness: split ContentView, fix Watch sync, add privacy manifest
- Split ContentView.swift (1268 lines) into platform-specific files:
  iOSContentView.swift, macOSContentView.swift, tvOSContentView.swift
- Replace broken App Group UserDefaults sync with WatchConnectivity
  (WCSession.updateApplicationContext) for reliable iPhone ↔ Watch communication
- Add PrivacyInfo.xcprivacy declaring crash/performance data (Sentry) and
  UserDefaults API usage
- Expand test suite from 1 to 44 tests covering WODTimerEngine
- Add error logging with [HealthKit], [Sound] prefixes
- Translate all Danish comments/strings to English; replace Sentry mentions
  in user-facing text with generic crash report language
- Remove unused App Group entitlements from iOS and Watch targets
- Add platformFilter=ios to Sentry-Dynamic to exclude it from Mac builds
- Fix duplicate sound file warnings (folder-sync vs explicit pbxproj entries)
- Fix Watch "Set Watch App Category" script running every build (add output path)
