+++ WODrounds/WODrounds.entitlements
@@ -4,9 +4,5 @@
 <dict>
 	<key>com.apple.developer.healthkit</key>
 	<true/>
-	<key>com.apple.security.application-groups</key>
-	<array>
-		<string>group.com.iamjarl.WODrounds</string>
-	</array>
 </dict>
 </plist>
