+++ WODrounds/da.lproj/Localizable.strings
@@ -19,5 +19,5 @@
 "Intervals" = "Intervaller";
 "About" = "Om";
 "Privacy" = "Privatlivspolitik";
-"No analytics or tracking. Crash reports are sent to Sentry to fix bugs." = "Ingen analyse eller sporing. Crashrapporter sendes til Sentry for at rette fejl.";
+"No analytics or tracking. Only crash reports are sent to help fix bugs." = "Ingen analyse eller sporing. Kun crashrapporter sendes for at rette fejl.";
 "No analytics or tracking." = "Ingen analyse eller sporing.";
